knowledge v2 done

+++ database/migrations/2019_09_19_174720_create_knowledge_table.php
@@ -25,9 +25,8 @@ class CreateKnowledgeTable extends Migration
             $table->integer('gpm')->nullable();
             $table->integer('xppm')->nullable();//xp per min
             $table->integer('lasthit')->nullable();
-            $table->decimal('hero_dmg', 5, 1)->nullable();
-            $table->decimal('tower_dmg', 5, 1)->nullable();
-
+            $table->integer('hero_dmg')->nullable();
+            $table->integer('tower_dmg')->nullable();
             $table->foreign('user_id')->references('id')->on('users')->onDelete('cascade');
             $table->timestamps();
         });
